web-platform_21_05_dev/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.6
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-grappelli==2.14.3
django-js-asset==1.2.2
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
openpyxl==3.0.7
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
